Replaced user list, and existing user search, with larger Zoom-wide search.
The recordings list is backed by the database, but the users list is still searching Zoom directly.
This means that if for some reason a recording is not present in the DB, you can still find it by
finding the user since the user's list of recordings is still being fetched from Zoom.  In fetching
from Zoom, it will also populate the database :)