Fixing incorrect xacml flavour issue.
This appears to be related to the order in which ingests are done.  If you
ingest files such that a call to _do_post occurs *without* files containing
something *after* you ingest a file containing a '+' character in its flavour
you will see it magically disappear the + from the next post.  The major issue
debugging this is that *this only disappears _after_ the post is sent*.  It is
*not* visible from the Python side, but it shows up in the debugger in Java.

I suspect this is the requests_toolbelt setting an incorrect content type,
since '+' renders as ' ' depending on the content type.  Further debugging required.