Fixing #62 and #63.  Let's talk about MySQL...
So, the relevant code here is comparing the Zoom UUID (09XO/8Y...) to an int
column in Recording.  This is something that (imo) should either break due to
mismatched datatypes, or return an empty set since it's *obviously* not a good
match.  Except it doesn't, it returns a seemingly random row.  Except it's not
truly random - it's doing a naive search character-by-character through the
string, returning the first matching ID.  In the case above, this was returning
the row with ID=9.  Casting the datatype to string so that proper string match
logic applies resolves the issue.  Why the database, or sqlalchemy, or python
itself isn't smart enough to detect and/or fix this is left as an exercise to
the reader.