Skip ingesting any episode ACL if --Blank ACL-- is selected
As --Blank ACL-- should be a dummy UI entry, no acl should be uploaded on selection. But currently an Deny-for-All policy will be created and uploaded to Opencast. This is obviously not the expected behaviour.
This patch will change that to skip episode ACL upload for --Blank ACL-- selected value.